에브리타임 공감 기능 개발 Rest docs 개발 완료

+++ src/main/java/com/project/everytimeclonecodingbackend/domain/post/service/PostService.java
@@ -33,7 +33,7 @@ public class PostService {
                 .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
 
         if (!member.isEmailAuthentication()) {
-            new CustomException(ErrorCode.NOT_CHECK_EMAIL);
+            throw new CustomException(ErrorCode.NOT_CHECK_EMAIL);
         }
 
         Post post = new Post(title, content, Category.valueOf(category), anonymous);
@@ -50,7 +50,7 @@ public class PostService {
                 .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
 
         if (!member.isEmailAuthentication() && !category.equals("자유게시판")) {
-            new CustomException(ErrorCode.NOT_CHECK_EMAIL);
+            throw new CustomException(ErrorCode.NOT_CHECK_EMAIL);
         }
 
         return postRepository.findAllByCategory(Category.valueOf(category), pageRequest);
@@ -63,7 +63,7 @@ public class PostService {
                 .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
 
         if (!member.isEmailAuthentication() && !category.equals("자유게시판")) {
-            new CustomException(ErrorCode.NOT_CHECK_EMAIL);
+            throw new CustomException(ErrorCode.NOT_CHECK_EMAIL);
         }
 
         return switch (tag) {
@@ -77,7 +77,7 @@ public class PostService {
         try {
             Category.valueOf(category);
         } catch (IllegalArgumentException e) {
-            new CustomException(ErrorCode.CATEGORY_NOT_FOUND);
+            throw new CustomException(ErrorCode.CATEGORY_NOT_FOUND);
         }
     }
 
@@ -89,7 +89,14 @@ public class PostService {
                 .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
 
         if (!member.isEmailAuthentication()) {
-            new CustomException(ErrorCode.NOT_CHECK_EMAIL);
+            throw new CustomException(ErrorCode.NOT_CHECK_EMAIL);
+        }
+
+        Post post = postRepository.findById(id)
+                .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
+
+        if (member.getId() != post.getMember().getId()) {
+            throw new CustomException(ErrorCode.POST_NOT_DELETE);
         }
 
         postRepository.deleteById(id);
@@ -105,7 +112,7 @@ public class PostService {
                 .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
 
         if (!member.isEmailAuthentication() && !post.getCategory().toString().equals("자유게시판")) {
-            new CustomException(ErrorCode.NOT_CHECK_EMAIL);
+            throw new CustomException(ErrorCode.NOT_CHECK_EMAIL);
         }
 
         return post;
@@ -119,10 +126,11 @@ public class PostService {
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new CustomException(ErrorCode.MEMBER_NOT_FOUND));
 
-        if (!member.isEmailAuthentication() && !post.getCategory().toString().equals("자유게시판")) {
-            new CustomException(ErrorCode.NOT_CHECK_EMAIL);
+        if (!member.isEmailAuthentication()) {
+            throw new CustomException(ErrorCode.NOT_CHECK_EMAIL);
         }
 
         return post.getMember().getId() == memberId;
     }
+
 }
